add sysctl's for panic on oom & reboot on panic
The use of panic_on_oom forces the kernel to throw a panic on oom when oom_killer has failed to resolve an oom situation. Then, kernel.panic forces the system to reboot on all panic events after X seconds instead of the system just hanging @ a panic waiting for intervention for a reboot.

The principle here, is a system locked up at a panic is useless, allow the kernel to get the system back online automatically. The caveat here is we need better tracking of reboot events with a top-list of daily/weekly reboots for sysops to investigate trending hw, software and similar issues that induce reboots (which is needed now anyways).